% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{Boot_mixtureVAR}
\alias{Boot_mixtureVAR}
\title{bootstrap mixtureVAR model}
\usage{
Boot_mixtureVAR(mod, nboot = 499, eps = 1e-08)
}
\arguments{
\item{mod}{the model to be bootstrapped}

\item{nboot}{number of bootstrap iterations}

\item{eps}{precision default to 1e-8}
}
\value{
return
}
\description{
Allows to perform a bootstrap on an estimated mixtureVAR model. This bootstrap is necessary for computing the impulse response functions.
}
