% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf.mixvar}
\alias{irf.mixvar}
\title{impulse response functions}
\usage{
\method{irf}{mixvar}(x, impulse = NULL, response = NULL,
  n.ahead = 10, ortho = TRUE, cumulative = FALSE, ci = 0.95,
  seed = NULL, ...)
}
\arguments{
\item{x}{an estimates mixturVAR model}

\item{impulse}{the names of the impulse variables}

\item{response}{the names of the response variables}

\item{n.ahead}{the number of timesteps going ahead}

\item{ortho}{TRUE/FALSE Should the residuals be orthogonalized}

\item{cumulative}{TRUE/FALSE Should a cumulative irf be used}

\item{ci}{nominal coverage rate of the interval}

\item{seed}{starting seed}
}
\value{
list
}
\description{
compute impulse response function for mixtureVARs
}
