% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnleffects.R
\name{mnlAveEffPlot2}
\alias{mnlAveEffPlot2}
\title{mnlAveEffPlot2}
\usage{
mnlAveEffPlot2(obj, varname, ColNames = NULL, R = 1500, nvals = 25,
  plot = TRUE, ...)
}
\arguments{
\item{obj}{mixvar}

\item{plot}{}
}
\description{
mnlAveEffPlot2
}
