% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mixvar}
\alias{plot.mixvar}
\title{plot mixtureVAR}
\usage{
plot.mixvar(x, plot.type = "multiple", names = NULL, main = NULL,
  sub = NULL, lty = NULL, lwd = NULL, col = NULL, ylim = NULL,
  ylab = NULL, xlab = NULL, nc, mar.multi = c(0, 4, 0, 4),
  oma.multi = c(6, 4, 6, 4), adj.mtext = NA, padj.mtext = NA,
  col.mtext = NA, path = NULL, ...)
}
\arguments{
\item{x}{irf.mixvar Object}
}
\description{
plot mixtureVAR
}
