% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plotBootIrf}
\alias{plotBootIrf}
\title{plot bootstrapped irf's of mixtureVAR}
\usage{
plotBootIrf(BOOTAUD, mod, impulse, response = "all", n.ahead = 20,
  bands = c(0.05, 0.95), type = c("mean", "median", "both", "p-value"),
  relation = "same", grey = FALSE, modnames = NULL)
}
\arguments{
\item{BOOTAUD}{bootstrap generated by Boot_FMVAR}

\item{mod}{the model to be bootstrapped}

\item{impulse}{which variable should be the impulse variable}

\item{n.ahead}{timepoint ahead for the irf}

\item{bands}{lower and upper quantile to be displayed}

\item{type}{type of plot used}
}
\description{
plot bootstrapped irf's of mixtureVAR
}
