% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plots2}
\alias{plots2}
\title{Generate plots for mixtureVAR}
\usage{
plots2(mod, type = "residuals", start = 1, frequency = 1)
}
\arguments{
\item{mod}{the estimated model by mixtureVAR}

\item{type}{type of plot desired}

\item{frequency}{}
}
\description{
Generate plots for mixtureVAR
}
